Move tests to tests/core/test_evaluation. Use agent test fixture.
